problems with converting numbers to cards, wip
