+++ src/main/kotlin/poker/model/simulation/OpponentScoreSimulator.kt
@@ -10,13 +10,12 @@ class OpponentScoreSimulator: HandScoreSimulator() {
         val tableCards = remainingTableCards(tableDetails, deck)
 
         val opponentCards = someOpponentCards(tableDetails, deck)
-        val opponentHands = opponentCards.map { it.plus(tableCards) }
+        val opponentHands = opponentCards.map { it.mergeCards(tableCards) }
         val opponentHandScores = opponentHands.map { handScoreConverter.convertToHandScore(it) }
 
         return opponentHandScores.maxByOrNull { it.value } ?: defaultHandScore
     }
 
-    // TODO - list of list doesn't read well, change to list of pairs or map of pairs?
     private fun someOpponentCards(tableDetails: TableDetails, deck: List<Card>): List<List<Card>> =
         (0 until tableDetails.totalPlayers).toList().map { opponentIndex ->
             val cardIndex = nextCardIndex(tableDetails, opponentIndex)
